+++ ChangeLog.md
@@ -1,5 +1,6 @@
 ## 1.2.0.0
 - Add `HasTypesUsing` and `HasTypesCustom` for custom traversals (Lysxia)
+- Improve type errors when no Generic instance is defined
 
 ### Breaking API changes
 - `HasType` now includes a reflexive case so that every type 'contains' itself (Matt Parsons)